GLASTONBURY/donor_wallet.py
from web3 import Web3
import json
import asyncio

# Team Instruction: Implement donor reputation wallet for API data contributions.
# Use Ethereum blockchain for transparency, inspired by Emeagwali’s decentralized vision.
class DonorWallet:
    def __init__(self):
        self.w3 = Web3(Web3.HTTPProvider("https://mainnet.infura.io/v3/YOUR_INFURA_KEY"))
        self.contract_address = "0x1234567890abcdef1234567890abcdef12345678"  # Mock contract
        with open("donor_wallet_abi.json", "r") as f:
            self.abi = json.load(f)
        self.contract = self.w3.eth.contract(address=self.contract_address, abi=self.abi)

    async def update_balance(self, wallet_id: str, export_codes: list) -> float:
        """Updates donor wallet balance based on API data contributions."""
        contribution = len(export_codes) * 0.0001  # 0.0001 ETH per code
        tx_hash = self.contract.functions.updateBalance(wallet_id, contribution).transact()
        receipt = self.w3.eth.wait_for_transaction_receipt(tx_hash)
        balance = self.contract.functions.getBalance(wallet_id).call()
        return balance / 1e18  # Convert wei to ETH

# Example usage
if __name__ == "__main__":
    wallet = DonorWallet()
    wallet_id = "eth:0x1234567890abcdef"
    export_codes = [99213, 99214, 93000]
    balance = asyncio.run(wallet.update_balance(wallet_id, export_codes))
    print(f"Updated wallet balance: {balance} ETH")