GLASTONBURY/neural_js_wrapper.py
import websocket
import json
import torch
import asyncio

# Team Instruction: Implement Neural JS/NeuroTS wrapper for real-time Neuralink integration.
# Use WebSocket for neural stream processing, inspired by Emeagwali’s real-time dataflow.
class NeuralJSWrapper:
    def __init__(self, neuralink_stream: str):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.neuralink_stream = neuralink_stream

    async def process_neural_stream(self, input_data: torch.Tensor) -> torch.Tensor:
        """Processes Neuralink data stream in real-time for API data enhancement."""
        ws = websocket.WebSocket()
        ws.connect(self.neuralink_stream)
        neural_data = json.loads(ws.recv())  # Mock neural data
        ws.close()

        neural_tensor = torch.tensor([neural_data.get("signal", 0)], dtype=torch.float32, device=self.device)
        combined_data = input_data + neural_tensor  # Simplified integration
        return combined_data

# Example usage
if __name__ == "__main__":
    wrapper = NeuralJSWrapper("wss://neuralink-api/stream")
    input_data = torch.ones(1000, dtype=torch.float32, device="cuda")
    result = asyncio.run(wrapper.process_neural_stream(input_data))
    print(f"Neural JS output shape: {result.shape}")