GLASTONBURY/requirements.txt
torch>=2.0.1
nvidia-cuda-runtime-cu12>=12.2
nvidia-cudnn-cu12>=8.9
nvidia-cuquantum>=23.03
numpy>=1.24.0
sqlalchemy>=2.0.0
modelcontextprotocol>=0.1.0
uvicorn>=0.18.0
fastapi>=0.95.0
qiskit>=0.43.0
qiskit-aer>=0.12.0
pycryptodome>=3.15.0
paho-mqtt>=1.6.1
gfortran>=11.0
libsdl2-dev>=2.0.20
websocket-client>=1.5.1
neurots>=0.1.0
web3.py>=6.0.0
eth-account>=0.8.0
aiohttp>=3.8.0
pyyaml>=6.0
jupyter>=1.0.0